Move all `*MetadataBlock` objects in `metadata/` folder
