+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/metadata/release_note_metadata_block.rb
